+++ modules/tools/cert-manager/r-cert-manager.tf
@@ -12,7 +12,7 @@ resource "helm_release" "cert-manager" {
   count      = var.enable_cert_manager ? 1 : 0
   name       = "cert-manager"
   chart      = "cert-manager"
-  repository = "https://charts.jetstack.io"
+  repository = var.cert_manager_chart_repository
   namespace  = kubernetes_namespace.cert-manager.0.metadata.0.name
   version    = var.cert_manager_chart_version
   dynamic "set" {