+++ modules/tools/kured/outputs.tf
@@ -1,0 +1,4 @@
+output "namespace" {
+  description = "Namespace used for Kured"
+  value       = local.namespace
+}