+++ outputs.tf
@@ -40,6 +40,14 @@ output "aks_user_managed_identity" {
   description = "The User Managed Identity used by AKS Agents"
 }
 
+##########################
+# AGIC outputs
+##########################
+output "agic_namespace" {
+  description = "Namespace used for AGIC"
+  value       = module.appgw.namespace
+}
+
 output "application_gateway_id" {
   description = "Id of the application gateway used by AKS"
   value       = module.appgw.application_gateway_id
@@ -58,4 +66,56 @@ output "public_ip_id" {
 output "public_ip_name" {
   value       = module.appgw.public_ip_name
   description = "Name of the public ip used by AKS application gateway"
+}
+
+##########################
+# AAD Pod Identity outputs
+##########################
+output "aad_pod_identity_namespace" {
+  description = "Namespace used for AAD Pod Identity"
+  value       = module.infra.aad_pod_identity_namespace
+}
+
+output "aad_pod_identity_azure_identity" {
+  description = "Identity object for AAD Pod Identity"
+  value       = module.infra.aad_pod_identity_azure_identity
+}
+
+##########################
+# Cert Manager outputs
+##########################
+output "cert_manager_namespace" {
+  description = "Namespace used for Cert Manager"
+  value       = module.certmanager.namespace
+}
+
+##########################
+# Velero outputs
+##########################
+output "kured_namespace" {
+  description = "Namespace used for Kured"
+  value       = module.kured.namespace
+}
+
+##########################
+# Velero outputs
+##########################
+output "velero_namespace" {
+  description = "Namespace used for Velero"
+  value       = module.velero.namespace
+}
+
+output "velero_storage_account" {
+  description = "Storage Account on which Velero data is stored."
+  value       = module.velero.storage_account
+}
+
+output "velero_storage_account_container" {
+  description = "Container in Storage Account on which Velero data is stored."
+  value       = module.velero.storage_account_container
+}
+
+output "velero_identity" {
+  description = "Azure Identity used for Velero pods"
+  value       = module.velero.velero_identity
 }